Added a CashPot class and begun remainder calc
CashPot class contains remainder calculator and pot total calculator
