+++ (iFound)ThisCoolSite/(iFound)ThisCoolSite.csproj
@@ -44,6 +44,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CashPot.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
